wait_for_server: ensure we sleep a bit when waiting for server
If we call wait_for_server() then self._SERVER_AGE is
the 'sleeping time' between 2 calls to 'get_server'. But by default
self._SERVER_AGE is 0, so we are hitting the Nova server hard.

This patch is to make sure that we wait for some time (2sec) if
self._SERVER_AGE has not been tweaked.

Change-Id: Ib9a93e922ff2e26f90d8d2e72f5cf6589054078a