Update load_balancer for v2 API
This patch updates the load_balancer.v2.load_balancer for the current
state of the API

It also updates the post_test_hook script to accept an argument to
restrict the functional tests run.  Since the load-balancer tests can
take a significant time to run, this allows us to run a parallel gate
job for the load-balancer functional tests.

Co-Authored-By: Michael Johnson <johnsomor@gmail.com>
Change-Id: I6c526cd8225a929d3d5984da10b6478312461e6f