Pass in app_name information to keystoneauth
keystoneauth (and os-client-config) support passing through app_name
and app_version. These affect the user-agent string. Allow our users
to set app_name and app_version.

So as not to cause an os-client-config minimum bump, check the arg list
of the occ constructor and only pass in app_name/app_version if it
accepts them. (This is informative and neat, but not, you know,
essential)

Change-Id: I515f93168adf35adc2521339255369b1ccfdfe0a