Transition nova security group tests to REST
novaclient v8 doesn't support these calls anymore. Transition them.

A few fixes in the non-test areas related to normalization. Also, we
allow users to pass in a dict to bypass server lookup, but then we
assume an object when we use it.

mocking the clients hides terible things. Yay requests_mock.

Change-Id: I39952800414058d7f8a77dfca7610bc8c1fcfd18