Futureproof keystone unit tests against new occ
We have a version of OCC coming out that fixes how keystoneclient is
constructed. The fix is good - however, it breaks a couple of places
where we're mocking to the old behavior. This will all go away once
we're done with keystoneclient-ectomy, so for now just put in a version
detection switch.

Change-Id: I9d4908ef12d40868dc207b130e3244577b9870e9