Include two transitive dependencies to work around conflicts
For people installing released shade from pip and then trying to use it
somehow through entrypoints, 1.20.0 is currently broken because of some
transitive dependencies, ordering issues and conflicting exclusion
ranges.

While this is by no means a comprehensive or sustainable solution to the
problem, it does unbreak the end users currently broken and is not
terribly onerous. Both of these additions are tracked in g-r and both
can be removed when the dependencies that pull in conflicting ranges of
things are removed.

Change-Id: I6a4a1ab5ab109f0650873201868e0f1c4d09c564