Fix create/rebuild tests to not need a ton of neutron
The tests are testing nova interactions, but because of _expand_server
they call neutron a bunch. This is hidden with neutronclient for some
reason - I'd look in to that more but we're deleting neutronclient
anyway.

To avoid us having to add a bunch of network mocks that are noise to the
tests, fake out has_service to return false for neutron to skip the
calls. We can then add them in on a case-by-case basis as needed.

Change-Id: Ia877232b8dc167a8414f1601ea82708ff5aa2be9