Port unversioned Version resources to resource2
Most services have a Version resource in their unversioned namespace,
generally predating the proxy interface. As such none of them are being
exposed or used as of right now, but rather than remove them we should
port them over as-is, then later figure out if we need to expose them
for use or if we don't actually need them.

Change-Id: I8f4ed67f6185bd1928ce73f8059d629dc0952f6b