Fix broken version discovery endpoints
In the version discovery spec, it is noted that both schema and netloc
on endpoints reported by version discovery documents are known to be
able to be broken, but shade was only fixing scheme. Update the code to
do the right thing. Also, add a test to ensure that this behavior occurs
correctly. While we're in the tests for this, have the discovery call
return a 300, which is what it returns. There are no bugs in this area,
but 300 is correct so we should test it that way.

Story: 2001027
Change-Id: Id04f6f033c9277e791b602f193d1eaf6ac73047f