Fix get_compute_usage normalization problem
When we normalize an empty usage report, start and end can be empty, so
the pop() call fails. Don't fail on that.

Also, stop forcing interface=admin in operator_cloud. We stopped adding
 this elsewhere already, because it turns out it's NOT actually a thing
 anyone should ever do by default. (It's an old keystone v2 thing -
jamielennox says it should be avoided like the plague) Make
operator_cloud consistent. (This will still honor interface=admin if
it's in the clouds.yaml file)

Change-Id: I432dc58d60fb56ce13332443435138354d66e9e1