Update test_user_update_password to overlay clouds.yaml
Rather than adding a bunch of v2/v3 logic, just use the demo cloud entry
in clouds.yaml but with the new user_id/password.

To make sure this works, go ahead and assign the member role to the
user, so that the project settings in the file will work too.

Change-Id: I1adc271d60c0ded25c639962d0974242009c7c9b