Avoid keystoneclient making yet another discovery call
Even with an endpoint_override set, keystoneclient will make one
discovery call if the keystoneclient.client.Client constructor is used.
Because we're not going to make that extra discovery call after we've
already made a discovery call when we do ksa calls, directly instantiate
the v2 or v3 objects ourselves. This lets us avoid the extra discovery
call so that we can remove it from our test fixtures. The discovery
call sequence now should be identical for ksa vs. ksc.

There is still an inefficiency here due to our discovery not using ksa
discovery. But we can optimize that later.

Change-Id: If7f8625745f80334b9a78d742dc4a9250a4eb72c