Don't fail hard on 404 from neutron FIP listing
We do a filtered FIP listing from _delete_servers, which has exposed our
list_floating_ips method as dying hard when it can't find a match.
That's pretty inconsistent with the rest of the API - return a [] in
that case but keep logging the big error message. The user should really
mark the cloud has floating_ip_source: None if they want the error
message to go away - which is indicated in the error message.

Change-Id: I76493e7cc004831c1a0620de9fbe004fdbc39e45