fix the bug that cannot create a listener by openstacksdk
When creating a listener with a loadbalance id by openstacksdk,there
have an error with the message "Listener must be created with a
loadbalancer or pool." I guess the reason is that I give the parameter
"loadbalancer_id"  but it be ignored in the code.

I modify the listener class, I add a new line load_balancer_id =
resource.Body('loadbalancer_id'), then successfully create a listener.

Change-Id: Id05fcdc0be8f50304a974c01d1aada4eaa4053cd
Closes-Bug: 1708780