Begin converting baremetal node tests
Shamelessly based upon pre-existing requests mock code in
shade, with the attempt to replace the existing baremetal
client mock testing.

This updates the way we're creating the ironic_client so that it
always does discovery, which _baremetal_client will always have
to do because of how the versioned discovery document is in ironic.

It also puts in a workaround for a bug that was found in keystoneauth's
new version discovery that we should eventually take out.

Change-Id: I0e811d69c3cba50f164cb2a3a1302b88fba51308