Handle glance image pagination links better
Glance returns pagination links like "/v2/images?marker=". keystoneauth
Adapter treats absolute links like that as if they are rooted on the
service url. Since the catalog has https://image.example.com/v2 in it,
the Adapter will be mounted on that, which means /v2/images will be
treated as "https://image.example.com/v2/v2/iamges - which is not
correct.

Doing a urljoin would also be incorrect in an OpenStack context, because
the service could be on a sub-url, such as https://example.com/image,
meaning the v2 API endpoint would be https://example.com/image/v2 -
which means a straight urljoin of https://example.com/image/v2 and
/v2/images?marker= would result in https://example.com/v2/images?marker=
which is wrong since it strips the /image path prefix.

The most correct things for glance to do would be to return a full
absolute url - or a full relative URL. Constructing the absolute url
shade-side is hard for the above reasons - we don't have the right
context.

Strip /v1/ or /v2/ from the front of the next links (yay!) to turn
it into a relative url - thereby letting adapter.get(endpoint) do the
correct thing.

While doing this, update the catalog fixture in the test suite to have a
trailing /v2 so that we can be sure we're doing the correct thing in our
unit tests. Also add suburl fixtures to ensure that we work with
suburls.

Change-Id: I6fc8484ed62a029a8ba8ff1d31a37ba69e3000cf