Remove ansible functional tests for now
These aren't actually testing openstacksdk in any way at the moment,
and having them is making us have a test-requirements depend on shade
which will make adding openstacksdk as a shade requirement a circular
dependency.

Remove them for now. To re-add them, we want to:

- add openstacksdk as a shade requirement
- update ansible to import openstack instead of import shade

Once that's done, we can add the tests back, but without adding shade as
an sdk test-requirement. That way we can have shade test that the
ansible modules work if someone installs shade and doesn't explicitly
install sdk, and we can have sdk test that the modules work if someone
installs sdk and not shade.

Change-Id: Icb26f9c066bad6c2c045ac949ac1864e26b5b837