Cleanup objects that we create on behalf of images
When the use performs create_image on a cloud that requires the image be
uploaded to swift, we upload the image content to swift for them. Once
the image is imported, the content in swift is no longer needed as
glance copies the data into its own area. Clean the images up so that
we're not just accumulating tons of content in swift.

Change-Id: Iede6441c0cad1d99e92fdfacf96d81aad479a93e