Start using Connection in openstack.cloud
Now that it's all one big happy tree, it's time to start using
Connection instead of direct Adapters in the old shade calls. This is a baby
step towards that.

The _munch_response + _get_and_munchify is clearly a terrible pattern.
We'll likely want to shift to using the actual SDK object layer, but
let's taking it a bit at a time.

This replaces all usage of _compute_client with _conn.compute. Amazingly
only one test broke, and that only broke because the test was broken and
removing an Exception rethrow caused it to be properly seen.

Change-Id: Ice60b2343ba30520bc27bef09ee9d5d7ad19411f