Fix py35 and pypy tox env
The basepython line was causing the py35 tests to actually run under
python2. Whoops.

It was doing the same to pypy. Add pypy to bindep with the pypy tag.
Cryptography and the version of pypi in xenial are incompatible, so
update the pypy job to run on fedora-26 instead, which has a new enough
pypy.

Shift the line so that it only applies to functional and ansible tests,
since that's where it's used.

Depends-On: I7c22e23b73ddfc18ee28e87d34d1988417b49ccb
Change-Id: Ia27a0aa0633dcf48c1bab06b3c803f996eae9efa