Add some docs about not using Profile
We list profile and authenticator as deprecated, but with no docs
providing suggestions as to what to do.

Add some docs.

These can be cleaned up a little bit as we continue to reorganize the
new code. For instance, you can actually do all of these just with the
Connection object, but only once we've released. Pointing people at
getting a cloud_config object for now allows writing code that will work
with both old and new interfaces.

Change-Id: I519c672d1a3905ef8695287d73fce0949576505d