Baremetal NIC list should return a list
Since the list_nics* methods were changed to use raw HTTP client rather
than the ironic client, they return a dict rather than a list. Instead
of getting this:

[{'address': '00:11:22:33:44:55', ...}, ...]

We get this:

{'ports': [{'address': '00:11:22:33:44:55', ...}, ...]}

This change removes this outer dict and returns to the old behaviour of
returning a list. This affects list_nics and list_nics_for_machine.

Change-Id: I3cb9ef5d97cf911cb4897b00ab4cef77b76efaa4