Allow to pass filters like domain to find_project/user
In multi-domain environments, project names or user names can be same
across domains. To retrieve a single project/user, we need to pass
domain information to find_project/find_user.

This commit defines _query_mapping to Project and User classes
and adds filtering parameters to find_project/find_user in _proxy.py.
The required field for filtering is 'domain_id' only, but it might be
useful to allow to pass other filtering fields.

Closes-Bug: #1739929
Needed-By: I8f59fa3f9b7c573485cd1572e5e9aae08f071e37
Change-Id: I60a8b3b83f6170b60d09c101b5c7035148283678