Prefer links dicts for pagination
The current pagination logic makes call to look for additional items
at times even if it's not needed, as it is only able to infer when to
make an additional call or not if the user is requesting a limit value
in their call.

However, calls to list() without limit don't do the right thing. If the
user does list() with no limit then list will always have to make an
unneeded additional call because it has no limit to compare against.

This additional call became evident in the next patch when attempting to
use sdk primitives in the shade layer and for applications like nodepool
would result in unacceptably higher traffic.

Luckily, there are mechanisms available that can be checked to see if
pagination is needed. Nova, Cinder and other Nova-like projects provide
a top level links dict named {resources_key}_links. The api-wg
recommendation is for a top level dict to be provided named 'links'.
Glance provides a top-level key named 'next'. Swift returns the total
number of objects in a container in a header.

The first three can be looked for with no additional options but simply
by inspecting the payload. The Swift case requires that we add an optional
"pagination_key" option to Resource.

Finally, we can keep the existing logic that applies when a limit was
given and the number of resources returned exactly matches the limit as
a heuristic for making an additional call for resources that do not
otherwise return next links. Since most resources that support
pagination should reutnr a next link, the fallback logic should only
rarely be triggered. If anyone encounters a situation with an
unacceptable additional call cost that stems from a service that does
not return next links, we can use that as motivation to get the service
in question fixed.

Change-Id: I7a015f05fb43b63805a5aa9f7e266069d5e52438