Port wait_for_ methods to use iterate_timeout
shade has a function called iterate_timeout which is used to, you
guessed it, iterate until a timeout is met. Update wait_for_status and
wait_for_deleted to use iterate_timeout.

While doing that, shift iterate_timeout to be in openstack.utils.

An attempt was made at shifting the parameter names of wait_for_ and
iterate_timeout to align, but aligning in both directions reads wrong.
wait_for_server with wait and interval reads well. iterate_timeout with
timeout and wait reads well. I think consistency is just going to have
to take a back seat to flow on this one.

Change-Id: Ida691638b57ff6d41b0a9e066a180b79b20642e9