Add function to make CloudRegion from session
In cases where a user already has a Session that they're happy with,
making a Connection to wrap it should be quite easy. The only settings
that are needed are Adapter level settings, and all of those have sane
defaults and don't actually need to be set.

Add a factory function that can be used in those cases.

Also, while we're in there, make CloudRegion use the word "region_name"
like everyone else.

This removes the get_region_name method that we added originally to occ in an
early attempt to support openstacksdk. However, nothing in the data
model makes sense if one service has one region_name and another service
has another region_name. Connection is a connection to a region of a
cloud.

Change-Id: I52fde4df188fc2747776c4b65bc7ae68c6e82f06