Remove legacy client factory functions
os-client-config provided helper functions for constructing legacy
python-*client objects, as well as Connection and OpenStackCloud
objects.

openstacksdk isn't in the business of constructing python-*client
objects for people, so let's remove these. We also dont' need helper
functions in the openstack.config space for constructing Connection
since we already have openstack.connect for that (and we'll fold
OpenStackCloud in soon enough)

Change-Id: I09e4fe258ff58a8e8aec5919369ff856a5102b76