Remove metric service
There is discussion on the mailing list about ceilometerclient removal
that started this.

The metric service is not an OpenStack service. It does not have an
entry in service-types-authority, nor does it have an API documented at
developer.openstack.org/api-ref.

That's not saying that someone couldn't install it in their OpenStack
cloud and that doing so is a bad idea. It's probably a great idea and
those people are likely quite happy.

Since this is openstacksdk we need to draw the line somewhere. The
add_service patch adds the ability for consumers to opt-in to
non-official services. Making an openstacksdk-metric package people
could use with a MetricService object and the proxy/resource objects
here would be a great idea.

As a follow-up we should potentially add the ability to add a list of
extra services to a clouds.yaml file.

Change-Id: Ib3810e3b11dbacc0215d486397e05370a2d7521e