Raise error when supplying invalid query params
Rather than stripping invalid query params and then
continuing the API call, lets actually throw an error
since we explicitly define what valid query params are.

This stops people from accidentally querying for an
incorrect param, getting back a list of 'everything' and
then acting on it assuming it was a valid query.

Adds a util function for getting keys from format strings
as the checking for invalid keys needs to take into account
what key may be required for the base_url.

Change-Id: If30badb4d71e521100a0e8974978eb6d5fa2699f