Remove meter service
As per the mailing list thread on ceilometerclient, the REST API for
ceilometer has been long deprecated and is now completely unsupported.

Remove the meter service and its sub-service alarm.

On the one hand, it's a REST API so there is no need for us to not
support it. As a general rule removing support for services that we have
already is a bad thing.

On the other hand, we haven't cut the 1.0 yet, and as ceilometer doesn't
exist anymore we can't really do devstack testing of it.

The previous patch adds support for out-of-tree services. So anyone who
cares about the 'meter' service can grab this, port it to Proxy2 and
then do an add_service.

Change-Id: Ic226c5ac048ba09e316c89cae1e8f2828e753e80