Migrate object_store to resource2/proxy2
This is the last thing we need to migrate before deleting the old
resource/proxy class.

There's a good pile of things related to uploading and downloading we
should port over from the openstack.cloud code. One of them, streaming
the object data, was added here. We can come back around and get all of
the large-object upload code supported.

Adds alias support into _BaseComponent much like prop used to have
because Swift needs it.

Port some of the unittests to requests-mock because it was just easier
to do that. Doing so uncovered some issues with case-sensitivity and
headers. As a result, there are some changes to resource2.Resource to
get case sensitivity sorted out. TODO comments have been left indicating
a few places for further cleanup, but those are internal and
non-essential.

The default value of has_body is changed to False for head calls.
Because. Well. Let's be honest. It's HEAD. There is no body. By
definition.

Change-Id: I8c4f18f78a77149e23b98f78af82b1d25ab7c4cf