Remove resource and proxy
Step one in the final removal of resource/proxy v1. Doing it as a
two commit chain just as a sanity check.

All of the services have been migrated to Resource2/Proxy2. Before we
rename those back to Resource/Proxy, delete the v1 versions and make
sure everything is still happy.

The next patch will do the renames.

Change-Id: I254a7474236ea4959db1bd00b397320b0ca27387