Change update/create method options
Rather than use booleans, switch to explicit definition of
which method is used. This is useful because 'update' can
when including Swift be 3 different methods, and that gets
messy with booleans. An explicit definition works nicer
and gives us more flexibility in future as well as a nicer
pattern for similar method remappings.

Depends-On: Icfa811571ec7c26617e0f8e9db086001fc0bc81f
Change-Id: I7e231a8c89ffd74fdd63aabafb408fadfdb69c67