Invert the attribute mapping
requests returns headers in a CaseInsensitiveDict because Headers are
case insensitive. Unfortunately, that only applies to keys, and our
current attribute mapping that we use to consume values from payloads is
{client_name: server_name}, which means we have to go through
shenanigans to get the comparison right.

Invert the mapping returned by _get_mapping and update the logic to
deal with the inversion.

While we're at it, remove _filter_component, as _consume_attrs already
does the same thing and handles the case sensitivity properly.

Change-Id: Ice38cc31635a364839472248f3814b0c8f65bfe1