Clean up a bit after the mapping inversion
The previous patch got the mapping inverted, but involved places where
CaseInsensitiveDict needed to be passed around. It's a quality of
Headers, so rework things to allow the dict type to be dictated by the
component class.

Add some helper methods that do common actions. These, and consume_attrs
can almost certainly be turned in to classmethods.

Remove calling filter_component or consume_attrs from limits - it's a
no-op. The Resource class handles this already.

Update the find_keypairs test. requires_id is a bit of a mess ...

Change-Id: Ic41f24fd70fb369370292c61a8867b31b8fd92ce