Update type conversion to handle subtypes more better
We have pretty much enough information in the Component declarations to
handle arbitrarily nested data descriptions. That's super cool. The only
thing we're missing is dealing with lists of declared types - and of
transparently handling sub-resources.

Resource objects want their construction parameters handed to them as
**kwargs arguments. So we need to figure out if the source data is a
dict and the target types is a Component, so we can pass the values to
the type constructor with **.

For lists, introduce a new Component parameter "list_type" which
declares the type of list in question. We could probably swing around
and make our own parameterizable List formatter and make this go away.
For now, this is easy enough.

The results can be seen with compute.v2.limits - the custom get method
can (almost completely) go away, since the type declaration can just
handle it for us.

Change-Id: Ieb5c3b60853000101504b82f1cd01b1a25370bf1