Make sure we use config defaults in profile compat code
The compat code to allow creating a Connection from a Profile has a
couple of bugs. First of all it doesn't return the config object it
creates. (whoops) Second, it doesn't build its config kwargs from a base
of the config defaults, so some code in CloudRegion breaks. The second
bug would also exist in from_session, although we haven't seen it in the
field yet.

Change-Id: Ibd0056e6484cfa9a4c8be6a8bc40a0eb2df76977