Add resource2/proxy2 wrappers with deprecations
Folks out there who may have resources subclassing resource2 will be
broken by the removal and rename back to resource. Add a resource2 and a
proxy2 that emits a deprecation warning from the constructor but
otherwise just subclasses resource and proxy.

While doing that, do the same thing for enable_logging which moved.

Change-Id: I76a3c9c3172942069b1e740189af530f35f41a1c