Update docs and pep8 tox environments
Make the docs tox environment use doc/requirements.

For pep8, just install the linters and skip installing the package.
Doing this but leaving in the constraints file keeps us in line with
upper-constraints on linters but lets our pep8 envs be tiny.

Finally, remove the doc8 line-length setting. It's not a big enough
difference to warrant changing it.

Change-Id: I3df81474fb52e2587d22c7789b6b553139eb37f2