Provide compatibility for people passing raw sessions
Code used to pass raw sessions to Resource methods. A common way to do
that was to do thins like FloatingIP.get(conn.session, name_or_id). As
conn.session is a session.Session that doesn't work anymore.

To help ease upgrade path issues, attach a reference to the Connection
into conn.session and so that we can pull the right adapter for a given
resource back out.

Add the neutron-grenade job to verify this works.

Change-Id: Ief9a0215ea2399b91d1d03a8048e73e6d7bedd64