Ensure Connection can be made from keyword arguments
Although driving configuration from clouds.yaml and environment
variables is super handy for many use cases, there are also plenty where
avoiding config files or environment variables is important.

Add a flag, "load_envvars" which defaults to True but can be disabled.

Update the Connection constructor to set load_yaml_config and
load_envvars to False if a named cloud is not given.

Update the docs to make it clear which form should be used for which use
case. Also, move the profile backwards compat documentation to its own
document so that it doesn't present needless complexity to users.

Change-Id: I6e05da5e73aff4143550e1d18fb0f743d51f2b70