Rename BaseProxy to Proxy
The base resource class is "Resource" but the base proxy class is
"BaseProxy" which makes reading the docs a bit strange. It was
originally not really a big real because nothing used Proxy directly.
But since we add Proxy instances to the Connection for every service in
service-types-authority which are now documented as part of the
Connection object, calling them BaseProxy is a bit weird.

Keep a BaseProxy object around so we don't break out of tree Proxy
objects, but add a deprecation to it so people can update.

Shift BaseProxy from the main user docs to the contributor layout docs.
Add showing inheritance, as well as docs for OpenStackSDKAdapter.

Change-Id: I2778b2a04ad0aed4bb6b217151f41503997db1c6