resource: don't early terminate list
APIs can set the max_limit to the number of items returned in each
query. If that max_limit is smaller than the limit given by the user,
the API will never return as many records as the limit, however that
doesn't mean there are no more records. So the list should not be
early terminated.

Closes-Bug: #1746410

Change-Id: I893f4b3ba0856476f81189d741de1aed8aa65e25