Add get_client_config method to CloudRegion
Client utilities sometimes need to be able to set some arbitrary
settings in their clouds.yaml file.

Add a get_client_config that takes an optional dictionary of defaults
and looks for a config section named for the client, falling back to
looking for a section named 'client'. (Thanks for the inspiration MySQL)

Change-Id: I9016746584d043f5b62170da6a52d4152885fe5b