Update base test case to use base from oslotest
We can get rid of our custom stdout/stderr handling by doing this. We
should push our 'attach-on-exception' code up to oslotest so we can get
rid of it locally.

Remove a direct mock of time.sleep. This is already handled in
the base test case. Also remove a time.sleep(40) - because of all of the
reasons.

Set the default for timeout to 3 seconds. Sometimes an error will cause
a requests-mock test to spin in an infinite loop. All of the unit tests
run in under a second and should stay that way.

Change-Id: I7f9ea357a85754a6d3f79955d3ffb2a68a0ff621